+++ pxr/base/gf/matrix4d.h
@@ -694,9 +694,9 @@ public:
     /// returns a different value type.
     GfVec3f TransformDir(const GfVec3f &vec) const {
         return GfVec3f(
-            vec[0] * _mtx[0][0] + vec[1] * _mtx[1][0] + vec[2] * _mtx[2][0],
-            vec[0] * _mtx[0][1] + vec[1] * _mtx[1][1] + vec[2] * _mtx[2][1],
-            vec[0] * _mtx[0][2] + vec[1] * _mtx[1][2] + vec[2] * _mtx[2][2]);
+            static_cast<float> (vec[0] * _mtx[0][0] + vec[1] * _mtx[1][0] + vec[2] * _mtx[2][0]),
+            static_cast<float> (vec[0] * _mtx[0][1] + vec[1] * _mtx[1][1] + vec[2] * _mtx[2][1]),
+            static_cast<float> (vec[0] * _mtx[0][2] + vec[1] * _mtx[1][2] + vec[2] * _mtx[2][2]));
     }
 
     /// Transforms the row vector \e vec by the matrix, returning the result.
@@ -716,9 +716,9 @@ public:
     /// (0, 0, 0, 1)).
     GfVec3f TransformAffine(const GfVec3f &vec) const {
         return GfVec3f(
-            vec[0] * _mtx[0][0] + vec[1] * _mtx[1][0] + vec[2] * _mtx[2][0] + _mtx[3][0],
-            vec[0] * _mtx[0][1] + vec[1] * _mtx[1][1] + vec[2] * _mtx[2][1] + _mtx[3][1],
-            vec[0] * _mtx[0][2] + vec[1] * _mtx[1][2] + vec[2] * _mtx[2][2] + _mtx[3][2]);
+            static_cast<float> (vec[0] * _mtx[0][0] + vec[1] * _mtx[1][0] + vec[2] * _mtx[2][0] + _mtx[3][0]),
+            static_cast<float> (vec[0] * _mtx[0][1] + vec[1] * _mtx[1][1] + vec[2] * _mtx[2][1] + _mtx[3][1]),
+            static_cast<float> (vec[0] * _mtx[0][2] + vec[1] * _mtx[1][2] + vec[2] * _mtx[2][2] + _mtx[3][2]));
     }
     /// @}
