+++ pxr/base/gf/vec2h.h
@@ -197,8 +197,8 @@ public:
 
     /// Multiplication by scalar.
     GfVec2h &operator*=(double s) {
-        _data[0] *= s;
-        _data[1] *= s;
+        _data[0] *= static_cast<float> (s);
+        _data[1] *= static_cast<float> (s);
         return *this;
     }
     GfVec2h operator*(double s) const {
@@ -261,7 +261,7 @@ public:
     /// \todo This was fixed for bug 67777. This is a gcc64 optimizer bug.
     /// By tickling the code, it no longer tries to write into
     /// an illegal memory address (in the code section of memory).
-    GfHalf Normalize(GfHalf eps = 0.001) {
+    GfHalf Normalize(GfHalf eps = 0.001f) {
         // TODO this seems suspect...  suggest dividing by length so long as
         // length is not zero.
         GfHalf length = GetLength();
@@ -269,7 +269,7 @@ public:
         return length;
     }
 
-    GfVec2h GetNormalized(GfHalf eps = 0.001) const {
+    GfVec2h GetNormalized(GfHalf eps = 0.001f) const {
         GfVec2h normalized(*this);
         normalized.Normalize(eps);
         return normalized;
@@ -296,8 +296,8 @@ PXR_NAMESPACE_OPEN_SCOPE
 inline
 GfVec2h::GfVec2h(class GfVec2d const &other)
 {
-    _data[0] = other[0];
-    _data[1] = other[1];
+    _data[0] = static_cast<float> (other[0]);
+    _data[1] = static_cast<float> (other[1]);
 }
 inline
 GfVec2h::GfVec2h(class GfVec2f const &other)
@@ -308,8 +308,8 @@ GfVec2h::GfVec2h(class GfVec2f const &other)
 inline
 GfVec2h::GfVec2h(class GfVec2i const &other)
 {
-    _data[0] = other[0];
-    _data[1] = other[1];
+    _data[0] = static_cast<float> (other[0]);
+    _data[1] = static_cast<float> (other[1]);
 }
 
 /// Returns component-wise multiplication of vectors \p v1 and \p v2.
@@ -348,7 +348,7 @@ GfGetLength(GfVec2h const &v)
 /// normalization. If the length of \c *v is smaller than \p eps then \c *v is
 /// set to \c *v/eps.  The original length of \c *v is returned.
 inline GfHalf
-GfNormalize(GfVec2h *v, GfHalf eps = 0.001)
+GfNormalize(GfVec2h *v, GfHalf eps = 0.001f)
 {
     return v->Normalize(eps);
 }
@@ -357,7 +357,7 @@ GfNormalize(GfVec2h *v, GfHalf eps = 0.001)
 /// If the length of this vector is smaller than \p eps, the vector divided by
 /// \p eps is returned.
 inline GfVec2h
-GfGetNormalized(GfVec2h const &v, GfHalf eps = 0.001)
+GfGetNormalized(GfVec2h const &v, GfHalf eps = 0.001f)
 {
     return v.GetNormalized(eps);
 }
