+++ pxr/base/gf/vec2i.h
@@ -185,8 +185,8 @@ public:
 
     /// Multiplication by scalar.
     GfVec2i &operator*=(double s) {
-        _data[0] *= s;
-        _data[1] *= s;
+        _data[0] *= static_cast<int> (s);
+        _data[1] *= static_cast<int> (s);
         return *this;
     }
     GfVec2i operator*(double s) const {
