+++ pxr/base/gf/vec3i.h
@@ -196,9 +196,9 @@ public:
 
     /// Multiplication by scalar.
     GfVec3i &operator*=(double s) {
-        _data[0] *= s;
-        _data[1] *= s;
-        _data[2] *= s;
+        _data[0] *= static_cast<int> (s);
+        _data[1] *= static_cast<int> (s);
+        _data[2] *= static_cast<int> (s);
         return *this;
     }
     GfVec3i operator*(double s) const {
