+++ pxr/usd/pcp/mapFunction.h
@@ -194,7 +194,7 @@ private:
         _Data() {};
 
         _Data(PathPair const *begin, PathPair const *end, bool hasRootIdentity)
-            : numPairs(end-begin)
+            : numPairs(static_cast<int> (end-begin))
             , hasRootIdentity(hasRootIdentity) {
             if (numPairs == 0)
                 return;
